requirements.txt
apturl==0.5.2
astroid==2.4.2
attrs==20.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.7.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dlib==19.21.1
dnspython==2.0.0
duplicity==0.8.12.0
email-validator==1.1.1
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fasteners==0.14.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-mongoengine==0.9.5
Flask-PyMongo==2.3.0
flask-talisman==0.7.0
Flask-WTF==0.14.3
future==0.18.2
httplib2==0.14.0
idna==2.10
iniconfig==1.0.1
isort==5.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.20.0
monotonic==1.5
more-itertools==8.5.0
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
packaging==20.4
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.2.0
pluggy==0.13.1
protobuf==3.6.1
py==1.9.0
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
pymongo==3.11.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==6.0.1
python-apt==2.0.0+ubuntu0.20.4.2
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.14.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
ssh-import-id==5.10
systemd-python==234
toml==0.10.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
usb-creator==0.3.7
wadllib==1.3.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
xkit==0.0.0
